+++ Woof/Woof/View/SitterFlow/SitterProfileView/SitterProfileView.swift
@@ -4,9 +4,6 @@ import SwiftUI
 struct SitterProfileView: View {
     // MARK: - Internal interface
 
-    /// View model responsible to manage data from model layer.
-    @StateObject var viewModel = SitterProfileViewModel()
-
     var body: some View {
         VStack {
             if isEditingMode {
@@ -18,6 +15,12 @@ struct SitterProfileView: View {
                         bio: $viewModel.bio,
                         pricePerHour: $viewModel.pricePerHour
                     )
+
+                    if viewModel.mandatoryFieldsAreEmpty {
+                        Text(mandatoryPlaceholderText)
+                            .padding(.vertical)
+                            .font(.system(.footnote))
+                    }
 
                     HStack {
                         Button(cancelButtonLabelText) {
@@ -32,7 +35,7 @@ struct SitterProfileView: View {
                                 await viewModel.save()
                             }
                         }
-                        .disabled(viewModel.name.isEmpty)
+                        .disabled(viewModel.mandatoryFieldsAreEmpty)
                     }
                 }
                 .padding()
@@ -50,8 +53,10 @@ struct SitterProfileView: View {
                         ratePerHour: viewModel.pricePerHour
                     )
 
-                    Button(editButtonLabelText) {
-                        isEditingMode.toggle()
+                    if !viewModel.sitterIsSet {
+                        Button(editButtonLabelText) {
+                            isEditingMode.toggle()
+                        }
                     }
                 }
                 .buttonStyle(CapsuleWithWhiteText())
@@ -94,11 +99,14 @@ struct SitterProfileView: View {
 
     // MARK: - Private interface
 
+    /// View model responsible to manage data from model layer.
+    @StateObject private var viewModel = SitterProfileViewModel()
     private let cancelButtonLabelText = "Cancel"
     private let saveButtonLabelText = "Save"
     private let editButtonLabelText = "Edit"
     private let tryAgainButtonLabelText = "Try Again"
     private let alertTitle = "Error"
+    private let mandatoryPlaceholderText = "Fields with * are mandatory"
     @State private var isEditingMode = false
 }